+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 0.7.2
+
+  * Added a new base backup and restore feature!
+    * This allows you to save your current base to the app's storage, and restore it and all of its building objects at another location, or in another game mode save.
+    * In base locations, a new menu item is added allowing you to store other people's bases.
+    * Only building objects from the vanilla game will be saved. This prevents the game from crashing for other players.
+    * This feature was made possible thanks to monkeyman192, who wrote the algorithm that makes base object position matrices compatible in new locations.
+  * Added username overriding.
+  * Fixed a bug causing the window to become invisible if DWM is disabled.
+  * Fixed a time stamp bug with locations.
+  * Fixed a bug that can cause the install directory to not be set by the user's preference.
+  * Fixed a bug causing the "Remove From Storage" option on selected locations appearing when the location is not stored.
+  * Fixed a bug that can cause the stored locations list to clear when switching modes from the menu and then teleporting.
+  * Fixed search queries getting reset prematurely.
+
 # 0.7.1
 
   * Added markdown support to descriptions.